+++ src/components/MultiHopTrade/components/MultiHopTradeConfirm/hooks/useAllowanceApproval.tsx
@@ -2,7 +2,7 @@ import { CONTRACT_INTERACTION } from '@shapeshiftoss/chain-adapters'
 import type { TradeQuoteStep } from '@shapeshiftoss/swapper'
 import { useCallback, useEffect } from 'react'
 import type { Hash } from 'viem'
-import { useApprovalTx } from 'components/MultiHopTrade/hooks/useAllowanceApproval/hooks/useApprovalTx'
+import { useApprovalTx } from 'components/MultiHopTrade/components/MultiHopTradeConfirm/hooks/useApprovalTx'
 import { useErrorHandler } from 'hooks/useErrorToast/useErrorToast'
 import { assertGetEvmChainAdapter, buildAndBroadcast } from 'lib/utils/evm'
 import { assertGetViemClient } from 'lib/viem-client'
